Refactor scene rendering
