+++ demo/src/models.js
@@ -1,0 +1,50 @@
+export default [
+  {
+    name: "assets/duck",
+    scale: [3, 3, 3],
+  },
+  {
+    name: "assets/airplane",
+    rotation: [Math.PI / 2, Math.PI, 0],
+    scale: [2, 2, 2],
+  },
+  {
+    name: "assets/skull",
+    rotation: [0, Math.PI, 0],
+  },
+  {
+    name: "assets/dragon",
+  },
+  {
+    name: "assets/boat",
+    rotation: [0, Math.PI, 0],
+  },
+  {
+    name: "assets/armadillo",
+    scale: [2, 2, 2],
+    rotation: [0, Math.PI, 0],
+  },
+  {
+    name: "assets/apricot",
+  },
+  {
+    name: "assets/frank",
+    scale: [0.05, 0.05, 0.05],
+  },
+  {
+    name: "assets/head",
+    scale: [2, 2, 2],
+  },
+  {
+    name: "assets/skull2",
+    scale: [3, 3, 3],
+  },
+  {
+    name: "assets/shiba",
+    scale: [5, 5, 5],
+  },
+  {
+    name: "assets/human",
+    scale: [0.05, 0.05, 0.05],
+  },
+];
